refactor: modularize store & client (#50)
* initial commit

* move config & errors to combined file
